moving hash function bugfix
