+++ core-algo/temp/graph.hpp
@@ -23,6 +23,25 @@ class VertexData{
         bool operator==(VertexData &v_data);
 };
 
+//overriding stds hash function for the VertexData class
+
+namespace std {
+  template <>
+  struct hash<VertexData>
+  {
+    std::size_t operator()(const VertexData& v_data) const
+    {
+        using std::size_t;
+        using std::hash;
+        // Compute individual hash values for first,
+        // second and combine them using XOR and bit shifting:
+        pair<double, double> geolocation = v_data.get_geolocation();
+        return (hash<double>()(geolocation.first) ^ (hash<double>()(geolocation.second) << 1));
+    }
+  };
+
+}
+
 class Vertex{
     private:
         int v_index;
