+++ core-algo/temp/vertexdata.cpp
@@ -6,25 +6,6 @@
 
 #include "graph.hpp"
 
-//overriding stds hash function for the VertexData class
-
-namespace std {
-  template <>
-  struct hash<VertexData>
-  {
-    std::size_t operator()(const VertexData& v_data) const
-    {
-        using std::size_t;
-        using std::hash;
-        // Compute individual hash values for first,
-        // second and combine them using XOR and bit shifting:
-        pair<double, double> geolocation = v_data.get_geolocation();
-        return (hash<double>()(geolocation.first) ^ (hash<double>()(geolocation.second) << 1));
-    }
-  };
-
-}
-
 // VertexData Class Implementation
 
 VertexData::VertexData(){
